+++ src/lib/outreach.test.ts
@@ -69,7 +69,6 @@ const providerSend = mock(
 mock.module("@/lib/resend", () => ({
   getResend: () => ({ emails: { send: providerSend } }),
   sendBoundedResendEmail: providerSend,
-  fetchReceivedResendEmail: async () => null,
   emailSender: (vertical?: string | null) =>
     vertical === "RESTAURANT"
       ? "Vincent from Restofrontapp <vincent@send.restofront.com>"
